requirements.txt
-i https://pypi.org/simple
anyio==3.6.1
click==8.1.3
ecdsa==0.18.0
fastapi==0.85.0
h11==0.14.0
idna==3.4
passlib==1.7.4
pyasn1==0.4.8
pydantic==1.10.2
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlalchemy==1.4.41
starlette==0.20.4
typing-extensions==4.4.0
uvicorn==0.18.3
